move duplicate function to test-helper
