+++ test-helper/src/lib.rs
@@ -12,6 +12,8 @@ pub mod starcoin_dao;
 pub mod txn;
 pub mod txpool;
 
+mod utils;
+
 pub use chain::{gen_blockchain_for_test, gen_blockchain_with_blocks_for_test};
 pub use dummy_network_service::DummyNetworkService;
 pub use network::{build_network, build_network_cluster, build_network_pair};
@@ -20,3 +22,4 @@ pub use starcoin_genesis::Genesis;
 pub use starcoin_node::NodeHandle;
 pub use starcoin_types::account::Account;
 pub use txpool::{start_txpool, start_txpool_with_miner, start_txpool_with_size};
+pub use utils::assert_that_version_control_has_no_unstaged_changes;
